refactor: Move functions functions to their own file
