+++ test/functions.ts
@@ -1,0 +1,7 @@
+import test from 'ava'
+
+test.todo('all_nf_releases')
+test.todo('lastest_stable_release_data')
+test.todo('release_data')
+test.todo('nextflow_bin_url')
+test.todo('install_nextflow')
