+++ test/main.ts
@@ -1,8 +1,8 @@
-import test from 'ava';
+import test from 'ava'
 
-const fn = () => 'foo';
+const fn = () => 'foo'
 
 test('fn() returns foo', t => {
   t.pass()
   // t.is(fn(), 'foo');
-});
+})
